moving queryBuilder to correct folder
